Fixing pandas and Clickhouse bugs. (#245)
* Avoid pandas groupbyapply bug.

* Allow manually triggering E2E workflow.

* Trigger E2E manually.

* Update secrets.

* Drop Clickhouse DBs after tests run.

* Use .env for modal secrets. Fix DB dropper.

* Run E2E for PRs with the 'Run E2E' label.

* Drop Redis keys in clear_test_dbs.py, add extra distributed debugging log messages.

* Fix clear_dbs documentation.

* Avoid Modal-inside-Modal.

* Avoid Modal-inside-Modal.

* Add missing wait back in.

+++ confirm/cloud/clickhouse.py
@@ -204,7 +204,9 @@ class Clickhouse:
     _results_table_exists: bool = False
 
     def __post_init__(self):
-        self.lock = redis.lock.Lock(self.redis_con, f"{self.job_id}:next_lock")
+        self.lock = redis.lock.Lock(
+            self.redis_con, f"{self.job_id}:next_lock", timeout=60
+        )
 
     @property
     def _host(self):
@@ -560,20 +562,24 @@ def get_ch_config(host=None, port=None, username=None, password=None, database=N
     )
 
 
-def clear_dbs(ch_client, redis_client, names=None, yes=False):
+def clear_dbs(
+    ch_client, redis_client, names=None, yes=False, drop_all_redis_keys=False
+):
     """
     DANGER, WARNING, ACHTUNG, PELIGRO:
         Don't run this function for our production Clickhouse server. That
         would be bad. There's a built-in safeguard to prevent this, but it's
         not foolproof.
 
-    Clear all databases (and database tables) from the Clickhouse server. This
-    should only work for our test database or for localhost.
+    Clear all databases (and database tables) from the Clickhouse and Redis
+    servers. This should only work for our test database or for localhost.
 
     Args:
-        client: Clickhouse client
+        ch_client: Clickhouse client
+        redis_client: Redis client
         names: default None, list of database names to drop. If None, drop all.
-        yes: bool, if True, don't ask for confirmation
+        yes: bool, if True, don't ask for confirmation before dropping.
+        drop_all_redis_keys: bool, if True, drop all Redis keys.
     """
     test_host = dotenv.dotenv_values()["CLICKHOUSE_TEST_HOST"]
     if not (
@@ -596,20 +602,32 @@ def clear_dbs(ch_client, redis_client, names=None, yes=False):
         to_drop = names
 
     if len(to_drop) == 0:
-        print("No databases to drop.")
-        return
+        print("No Clickhouse databases to drop.")
+    else:
+        print("Dropping the following databases:")
+        print(to_drop)
+        if not yes:
+            print("Are you sure? [yN]")
+            yes = input() == "y"
 
-    print("Dropping the following databases:")
-    print(to_drop)
-    if not yes:
-        print("Are you sure? [yN]")
-        yes = input() == "y"
+        if yes:
+            for db in to_drop:
+                cmd = f"drop database {db}"
+                print(cmd)
+                ch_client.command(cmd)
+                if redis_client is not None:
+                    keys = list(redis_client.scan_iter("*{db}*")) + list(
+                        redis_client.scan_iter("{db}*")
+                    )
+                    print("deleting redis keys", keys)
+                    redis_client.delete(*keys)
+                else:
+                    print("no redis client, skipping redis keys")
 
-    if yes:
-        for db in to_drop:
-            cmd = f"drop database {db}"
-            print(cmd)
-            ch_client.command(cmd)
-            keys = list(redis_client.scan_iter("*{db}*"))
-            print("deleting redis keys", keys)
-            redis_client.delete(*keys)
+    if drop_all_redis_keys:
+        keys = list(redis_client.scan_iter("*"))
+        print("drop_all_redis_keys=True, deleting redis keys", keys)
+        if not yes:
+            print("Are you sure? [yN]")
+            yes = input() == "y"
+        redis_client.delete(*keys)
